flask_backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
blinker==1.9.0
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
eval_type_backport==0.2.2
faiss-cpu==1.12.0
Flask==3.1.2
flask-cors==6.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mistralai==1.9.10
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
